// app/interfaces/MainComponent/IOverviewSectionComponent.ts
export interface IOverviewSection {
    _id: string;
    id?: Number;
    section_title?: string;
    section_discription?: string;
    image1?: string;
    image2?: string;
    image3?: string;
    image4?: string;
    image5?: string;
    image6?: string;
    image7?: string;
    image8?: string;
    image9?: string;
    image10?: string;
    heading1?: string;
    heading2?: string;
    heading3?: string;
    heading4?: string;
    heading5?: string;
    heading6?: string;
    heading7?: string;
    heading8?: string;
    heading9?: string;
    heading10?: string;
    description1?: string;
    description2?: string;
    description3?: string;
    description4?: string;
    description5?: string;
    description6?: string;
    description7?: string;
    description8?: string;
    description9?: string;
    description10?: string;
}

export interface IOverviewSectionInput {
    id?: Number;
    section_title?: string;
    section_discription?: string;
    image1?: string;
    image2?: string;
    image3?: string;
    image4?: string;
    image5?: string;
    image6?: string;
    image7?: string;
    image8?: string;
    image9?: string;
    image10?: string;
    heading1?: string;
    heading2?: string;
    heading3?: string;
    heading4?: string;
    heading5?: string;
    heading6?: string;
    heading7?: string;
    heading8?: string;
    heading9?: string;
    heading10?: string;
    description1?: string;
    description2?: string;
    description3?: string;
    description4?: string;
    description5?: string;
    description6?: string;
    description7?: string;
    description8?: string;
    description9?: string;
    description10?: string;
}